Refactor: 트랙 자동 배치 로직 usePlaceRails 훅으로 분리
